refactor(graph-ui): split App.jsx into smaller panels and hooks
Break the 1355-line App.jsx monolith into focused modules with clear
ownership boundaries:

Components (src/components/):
  - Pill, Section, ContextMenu, EdgeDialog, FileInputButton
  - LeftPanel, GraphCanvas, RightPanel

Hooks (src/hooks/):
  - useCytoscape - Cytoscape lifecycle, events, and selection sync
  - useUndoRedo - undo/redo state and keyboard shortcut

Utilities (src/lib/):
  - boot.js - getBootConfig extracted
  - file-utils.js - readFileText, readFileBase64 extracted

No behavior or visual output changes. All state remains in App.
